requirements.txt
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.31
langchain-openai==0.0.3
transformers==4.30.2
langchain-text-splitters==0.0.1
sentence-transformers==2.2.2
huggingface-hub==0.16.4
openai>=1.6.1,<2.0.0
faiss-cpu==1.10.0
pypdf==5.4.0
python-dotenv==1.1.0
python-docx==1.1.0
docx2txt==0.9
torch==2.6.0
numpy==1.26.4
aiofiles==23.2.1
aiohttp==3.9.3
tenacity==8.2.3
nest_asyncio==1.6.0
aiogram==3.19.0
psycopg2-binary==2.9.10
aiosqlite==0.19.0
python-telegram-bot==22.1